Fix local asset loading issue
Turns out deep within BabylonJS, an XmlHttpRequest returns status code of 0 for local assets as opposed to status code of 200, so foreign objects never load!
So we add our own fork of the babylon dist code here. Also rejig the scene code, so that we add the habitat model when the Build Habitat button is pressed