Implement MVC architecture
Refactoring the long procedural nature of the scene setup into
different classes. The idea is to establish an architecture similar
to Model - View - Controller. This commit sets up the concept of the
Controller and Model/Data layer (called SceneGraph). So the idea is
that the controller accepts UserInput and will update the underlying
data layer ie the scene graph. There's no real view layer yet (as the
BabylonJS rendering functionality is hard to extract into a separate
rendering layer), but in future this might be added in the form of
some UI layer which informs about resources...