Add concept of Quests
The adds the notion of quests, which is an array of missions to
fulfill. Quests have a boolean condition that has to be met and
a reward which the SceneController has to interpret. As a side
effect it gives a mechanism for a win condition. Right now the last
quest basically means the player wins the game.